+++ src/main/java/nl/han/minor/alliander/rfid/prototype/persistence/interfaces/ITagDatabase.java
@@ -1,5 +1,6 @@
 package nl.han.minor.alliander.rfid.prototype.persistence.interfaces;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import nl.han.minor.alliander.rfid.prototype.persistence.DAOs.TagDAO;
@@ -7,5 +8,5 @@ import nl.han.minor.alliander.rfid.prototype.persistence.DAOs.TagDAO;
 public interface ITagDatabase {
   public List<TagDAO> getAllTags();
 
-  public TagDAO getTagsFromID(long id);
+  public TagDAO getTagsFromID(BigInteger id);
 }
